requirements.txt
django
psycopg2-binary
django-rest-framework